created a MasterDataLayout to allow nested layout options
