+++ ui/src/components/layouts/MasterDataLayout.tsx
@@ -1,20 +1,9 @@
-import type { Theme } from "@mui/material";
-import { Box, Stack, Typography } from "@mui/material";
-import { useMediaQuery } from "@mui/material";
-import Divider from "@mui/material/Divider";
-import Drawer from "@mui/material/Drawer";
-import List from "@mui/material/List";
-import ListItem from "@mui/material/ListItem";
-import ListItemButton from "@mui/material/ListItemButton";
-import ListItemIcon from "@mui/material/ListItemIcon";
-import ListSubheader from "@mui/material/ListSubheader";
+import { Box } from "@mui/material";
 import { useQuery } from "@tanstack/react-query";
 import { gql, request } from "graphql-request";
-import Image from "next/image";
 import * as React from "react";
 import { ReactElement, useEffect, useMemo, useState } from "react";
 
-import Link from "@/components/Link";
 import { DataProvider } from "@/context/DataContext";
 import {
   Detection,
@@ -23,45 +12,7 @@ import {
   useFeedsQuery,
 } from "@/graphql/generated";
 import { Candidate } from "@/pages/moderator/candidates";
-import wordmark from "@/public/wordmark/wordmark-white.svg";
 import { AIData } from "@/types/DataTypes";
-import { analytics } from "@/utils/analytics";
-
-import { TopNav } from "./Devias-dashboard/vertical-layout/top-nav";
-import navigationList from "./navigationList";
-
-// const drawerWidth = 240;
-const drawerWidth = "280px";
-
-function Brand({ onClick }: { onClick?: () => void }) {
-  return (
-    <Typography variant="h6" noWrap overflow="visible">
-      <Link
-        href="/"
-        color="inherit"
-        underline="none"
-        sx={{
-          height: "100%",
-          display: "flex",
-          alignItems: "center",
-          justifyContent: "center",
-        }}
-        onClick={() => {
-          if (onClick) onClick();
-          analytics.nav.logoClicked();
-        }}
-      >
-        <Image
-          src={wordmark.src}
-          alt="Orcasound"
-          width={140}
-          height={60}
-          priority={true}
-        />
-      </Link>
-    </Typography>
-  );
-}
 
 const endpointOrcahello =
   "https://aifororcasdetections.azurewebsites.net/api/detections?";
@@ -91,16 +42,12 @@ const standardizeFeedName = (name: string) => {
   switch (name) {
     case "Beach Camp at Sunset Bay":
       return "Sunset Bay";
-      break;
     case "North SJC":
       return "North San Juan Channel";
-      break;
     case "Haro Strait":
       return "Orcasound Lab";
-      break;
     default:
       return name;
-      break;
   }
 };
 const lookupFeedName = (id: string, feedList: Feed[]) => {
@@ -124,9 +71,8 @@ const lookupFeedId = (name: string, feedList: Feed[]) => {
   return id;
 };
 
-export function ModeratorLayout({ children }: { children: React.ReactNode }) {
+export function MasterDataLayout({ children }: { children: React.ReactNode }) {
   //// DATA
-  const lgUp = useMediaQuery((theme: Theme) => theme.breakpoints.up("lg"));
 
   const [nowPlaying, setNowPlaying] = useState({} as Candidate);
 
@@ -320,135 +266,6 @@ export function ModeratorLayout({ children }: { children: React.ReactNode }) {
     };
   }, [datasetHuman, aiDetections, feedList, isSuccess, nowPlaying]);
 
-  //// COMPONENTS
-
-  const listItem = (
-    title: string,
-    path: string,
-    icon: ReactElement,
-    key: string,
-  ) => (
-    <Link
-      key={key}
-      href={path}
-      underline="none"
-      width={0.9}
-      sx={{
-        color: "inherit",
-        opacity: ".75",
-        ["&.active"]: {
-          background: "rgba(255,255,255,.15)",
-          opacity: "1",
-          // background: "#258DAD",
-        },
-      }}
-    >
-      <ListItem
-        disablePadding
-        sx={{
-          background: "inherit",
-          borderRadius: "8px",
-        }}
-      >
-        <ListItemButton style={{ padding: "14px 20px" }}>
-          <ListItemIcon
-            style={{
-              color: "inherit",
-              minWidth: "48px",
-              fontSize: "27px",
-              opacity: ".75",
-            }}
-          >
-            {icon}
-          </ListItemIcon>
-          <Typography
-            sx={{
-              margin: 0,
-              fontWeight: "inherit",
-              fontFamily: "inherit",
-              fontSize: "inherit",
-            }}
-          >
-            {title}
-          </Typography>
-        </ListItemButton>
-      </ListItem>
-    </Link>
-  );
-
-  const subheader = (content: string) => (
-    <ListSubheader
-      component="div"
-      id={content.replace(" ", "-").toLowerCase()}
-      sx={{
-        background: "transparent",
-        color: "inherit",
-        textTransform: "uppercase",
-        fontWeight: "bold",
-        opacity: ".75",
-        lineHeight: 2.5,
-        marginTop: "1.2rem",
-        fontSize: "14px",
-      }}
-    >
-      {content}
-    </ListSubheader>
-  );
-
-  interface NavDiv {
-    title?: string;
-    kind: string;
-    children?: NavItem[];
-  }
-
-  interface NavItem {
-    title: string;
-    path: string;
-    icon: ReactElement;
-  }
-
-  const navDiv = (div: NavDiv, index: number) => {
-    let component;
-    switch (div.kind) {
-      case "divider":
-        component = <Divider key={index} />;
-        break;
-      case "subheader":
-        component = (
-          <List
-            key={index}
-            component="nav"
-            aria-labelledby="nested-list-subheader"
-            subheader={div.title && subheader(div.title)}
-            sx={{
-              background: "transparent",
-              color: "inherit",
-              fontFamily: "Mukta, Montserrat, sans-serif",
-              fontWeight: "bold",
-              fontSize: "18px",
-            }}
-          >
-            {div.children &&
-              div.children.map((item, index) =>
-                listItem(
-                  item.title,
-                  item.path,
-                  item.icon,
-                  `${item.title}-${index}`,
-                ),
-              )}
-          </List>
-        );
-    }
-    return component;
-  };
-
-  const DrawerList = (
-    <Box sx={{ overflow: "auto", marginTop: "5px" }}>
-      {navigationList.map((item, index) => navDiv(item, index))}
-    </Box>
-  );
-
   //// RENDER
 
   return (
@@ -465,7 +282,6 @@ export function ModeratorLayout({ children }: { children: React.ReactNode }) {
         flexDirection: "column",
       }}
     >
-      {/* <Header /> */}
       {process.env.NODE_ENV === "development" && (
         <button
           onClick={() => setUseLiveData((prev) => !prev)}
@@ -480,91 +296,12 @@ export function ModeratorLayout({ children }: { children: React.ReactNode }) {
           {useLiveData ? "Using LIVE data" : "Using SEED data"}
         </button>
       )}
-      <TopNav />
-      <Box sx={{ flexGrow: 1, display: "flex", width: "100vw" }}>
-        {lgUp && (
-          <Drawer
-            anchor="left"
-            open
-            sx={{ width: drawerWidth }}
-            PaperProps={{
-              sx: {
-                width: drawerWidth,
-                backgroundColor: "base.main",
-                color: "base.contrastText",
-              },
-            }}
-            variant="permanent"
-          >
-            <Stack sx={{ height: "100%" }}>
-              <Stack
-                alignItems="center"
-                direction="row"
-                spacing={2}
-                sx={{ padding: "4px 36px" }}
-              >
-                <Brand />
-              </Stack>
-              <Stack
-                component="nav"
-                spacing={2}
-                sx={{
-                  flexGrow: 1,
-                  px: 2,
-                }}
-              >
-                {DrawerList}
-              </Stack>
-            </Stack>
-          </Drawer>
-        )}
 
-        {/* <div key={"left"} className="drawer-div">
-          <Drawer
-            PaperProps={{
-              sx: {
-                backgroundColor: "base.main",
-                color: "base.contrastText",
-              }
-            }}
-            variant="permanent"
-            sx={{
-              flexShrink: 0,
-              width: drawerWidth,
-              [`& .MuiDrawer-paper`]: {
-                width: drawerWidth,
-                boxSizing: "border-box",
-              },
-            }}
-          >
-          <Box 
-            sx={{
-              backgroundColor: "base.main",
-              color: "base.contrastText",
-              height: "100vh",
-              // overflow: "scroll",
-              padding: "0px 16px",
-              fontWeight: "700"
-            }}
-          >
-            <div style={{height: ".5rem"}} />
-            <Toolbar />
-            {DrawerList}
-          </Box>
-          </Drawer>
-          </div> */}
-        <Box
-          // maxWidth="xl"
-          sx={{ width: "100%", padding: 0, margin: 0 }}
-        >
-          <DataProvider data={dataset}>{children}</DataProvider>
-        </Box>
-        {/* <PlayBar candidate={nowPlaying} /> */}
-      </Box>
+      <DataProvider data={dataset}>{children}</DataProvider>
     </Box>
   );
 }
 
-export function getModeratorLayout(page: ReactElement) {
-  return <ModeratorLayout>{page}</ModeratorLayout>;
+export function getMasterDataLayout(page: ReactElement) {
+  return <MasterDataLayout>{page}</MasterDataLayout>;
 }
